fix: menu animation
